+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+<a name="2.59.0"></a>
+
+# [2.59.0](https://gitlab.com/atjdev/ezio-api/-/merge_requests/301) (2022-09-23)
+
+### Features
+
+- [**M017**] **fdp**: add column for AGT approval.
+
+<a name="2.58.0"></a>
+
+# [2.58.0](https://gitlab.com/atjdev/ezio-api/-/merge_requests/299) (2022-09-21)
+
+### Features
+
+- [**M019**] **gis**: add column for AT2.
+
+<a name="2.57.0"></a>
+
+# [2.57.0](https://gitlab.com/atjdev/ezio-api/-/merge_requests/293) (2022-09-15)
+
+### Features
+
+- [**G002**] **gap**: feature pesanan seragam.
+
 <a name="2.56.0"></a>
 
 # [2.56.0](https://gitlab.com/atjdev/ezio-api/-/merge_requests/296) (2022-09-15)